+++ src/ConsulSharp/V1/ACL/Models/ReplicationStatusModel.cs
@@ -36,7 +36,7 @@ namespace ConsulSharp.V1.ACL.Models
         /// After the initial sync, replica lag should be on the order of about 30 seconds.
         /// </summary>
         [JsonProperty("ReplicatedIndex")]
-        public long ReplicatedIndex { get; set; }
+        public string ReplicatedIndex { get; set; }
 
         /// <summary>
         /// The UTC time of the last successful sync operation.