+++ src/ConsulSharp/V1/ACL/Models/TokenModel.cs
@@ -11,12 +11,12 @@ namespace ConsulSharp.V1.ACL.Models
         /// The create index.
         /// </summary>
         [JsonProperty("CreateIndex")]
-        public long CreateIndex { get; set; }
+        public string CreateIndex { get; set; }
 
         /// <summary>
         /// The modify index.
         /// </summary>
         [JsonProperty("ModifyIndex")]
-        public long ModifyIndex { get; set; }
+        public string ModifyIndex { get; set; }
     }
 }