+++ src/ConsulSharp/V1/ConsulClientV1.cs
@@ -2,6 +2,7 @@
 using ConsulSharp.V1.ACL;
 using ConsulSharp.V1.Event;
 using ConsulSharp.V1.KeyValue;
+using ConsulSharp.V1.Snapshot;
 using ConsulSharp.V1.Status;
 
 namespace ConsulSharp.V1
@@ -13,6 +14,7 @@ namespace ConsulSharp.V1
             ACL = new ACLProvider(polymath);
             Event = new EventProvider(polymath);
             KeyValue = new KeyValueProvider(polymath);
+            Snapshot = new SnapshotProvider(polymath);
             Status = new StatusProvider(polymath);
         }
 
@@ -21,6 +23,8 @@ namespace ConsulSharp.V1
         public IEvent Event { get; }
 
         public IKeyValue KeyValue { get; }
+
+        public ISnapshot Snapshot { get; }
 
         public IStatus Status { get; }
     }