+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 **FEATURES:**
 
  * All ACL APIs - Create, Update, Read and Delete Token.
- * All KeyValue APIs - Write, Read, Delete Value.
- * All Status APIs - Get leader and Raft peers.
  * All Event APIs - Fire and list events.
+ * All KeyValue APIs - Write, Read, Delete Value.
+ * All Snapshot APIs - Generate and Restore.
+ * All Status APIs - Get leader and Raft peers.